+++ CHANGELOG.md
@@ -1,5 +1,17 @@
+# 2.0.1 (2025-11-11)
+
+> This is a patch release for version 2.0.0 and contains breaking changes
+
+- Migrated session storage from hive_ce to flutter_secure_storage, after this related methods and getters are now asynchronous
+- Deprecated Session related methods and getters. Use Token related methods and getters instead (will be removed in future releases)
+- Added AES GCM encryption for LocalStorage
+- By default, LocalStorage now uses AES GCM encryption. You can add your own encryption cipher by passing customCipher while initializing LocalStorage
+- Added SecureStorage class for secure key-value storage using flutter_secure_storage
+
 # 2.0.0
+
 > This release contains breaking changes
+
 - Make LocalStorage singleton and renamed getInstance with initialize, use LocalStorage.instance or LocalStorage.i or LocalStorage() for accessing member functions and getters
 - Migrated to hive_ce
 - Changed dart min sdk to 3.8